refactor(bot): renames files + updates folder archi
