+++ bot/src/commands/_index.ts
