+++ bot/src/constants/emojis.ts
