+++ bot/src/constants/validations.ts
@@ -1,1 +1,1 @@
-export const SCHEDULED_MESSAGE_TAG_REGEX = /\$(.*?)\$/g
+export const SCHEDULED_MESSAGE_TAGS_REGEX = /\$(.*?)\$/g;
