+++ bot/src/helpers/dates/get-day-of-next-wee-timestamp.ts
