+++ bot/src/jobs/enrollment-message/message-tags-map.ts
@@ -1,0 +1,1 @@
+export const MESSAGE_TAGS_MAP = {};
